Move decorators into their own module.
